Rename Object to CompositionObject (Fixes #81)
