+++ tests/BlockElement/CheckboxesUnitTest.php
@@ -3,7 +3,7 @@ namespace Slack\Tests\BlockElement;
 
 use Maknz\Slack\BlockElement\Checkboxes;
 use Maknz\Slack\BlockElement\Text;
-use Maknz\Slack\Object\Option;
+use Maknz\Slack\CompositionObject\Option;
 use Slack\Tests\TestCase;
 
 class CheckboxesUnitTest extends TestCase
