+++ tests/BlockElement/RadioButtonsUnitTest.php
@@ -4,7 +4,7 @@ namespace Slack\Tests\BlockElement;
 use InvalidArgumentException;
 use Maknz\Slack\BlockElement\RadioButtons;
 use Maknz\Slack\BlockElement\Text;
-use Maknz\Slack\Object\Option;
+use Maknz\Slack\CompositionObject\Option;
 use Slack\Tests\TestCase;
 
 class RadioButtonsUnitTest extends TestCase
